path compilation issues tests moved to proper location
